+++ openrouteservice-api-tests/src/test/java/heigit/ors/services/routing/ResultTest.java
@@ -53,6 +53,9 @@ public class ResultTest extends ServiceTest {
 		addParameter("preference", "fastest");
 		addParameter("bikeProfile", "cycling-regular");
 		addParameter("carProfile", "driving-car");
+
+		// query for testing the alternative routes algorithm
+        addParameter("coordinatesAR", "8.680401,49.437436|8.746362,49.414191");
 	}
 
     @Test
@@ -1581,4 +1584,26 @@ public class ResultTest extends ServiceTest {
                 .body("features[0].geometry.coordinates.size()", is(34))
                 .statusCode(200);
 	}
+
+    @Test
+    public void testAlternativeRoutes() {
+        given()
+                .param("coordinates", getParameter("coordinatesAR"))
+                .param("instructions", "true")
+                .param("preference", getParameter("preference"))
+                .param("profile", getParameter("carProfile"))
+                .param("options", "{\"alternative_routes_count\": 2}")
+                .when().log().ifValidationFails()
+                .get(getEndPointName())
+                .then()
+                .assertThat()
+                .body("any { it.key == 'routes' }", is(true))
+                .body("routes.size()", is(2))
+                .body("routes[0].summary.distance", is(8178.2f))
+                .body("routes[0].summary.duration", is(1087.4f))
+                .body("routes[1].summary.distance", is(10670.9f))
+                .body("routes[1].summary.duration", is(1414))
+                .statusCode(200);
+    }
 }
+