+++ openrouteservice/src/main/java/heigit/ors/api/requests/routing/RouteRequestHandler.java
@@ -58,7 +58,7 @@ public class RouteRequestHandler extends GenericHandler {
         this.errorCodes.put("UNKNOWN_PARAMETER", RoutingErrorCodes.UNKNOWN_PARAMETER);
     }
 
-    public  RouteResult generateRouteFromRequest(RouteRequest request) throws StatusCodeException{
+    public RouteResult[] generateRouteFromRequest(RouteRequest request) throws StatusCodeException{
         RoutingRequest routingRequest = convertRouteRequest(request);
 
         try {
@@ -166,6 +166,19 @@ public class RouteRequestHandler extends GenericHandler {
             params = processRouteRequestOptions(request, params);
         }
 
+        if (request.hasAlternativeRoutes()) {
+            if (request.getCoordinates().size() > 2) {
+                throw new IncompatibleParameterException(RoutingErrorCodes.INCOMPATIBLE_PARAMETERS, RouteRequest.PARAM_ALTERNATIVE_ROUTES, "(number of waypoints > 2)");
+            }
+            RouteRequestAlternativeRoutes alternativeRoutes = request.getAlternativeRoutes();
+            if (alternativeRoutes.hasTargetCount())
+                params.setAlternativeRoutesCount(alternativeRoutes.getTargetCount());
+            if (alternativeRoutes.hasWeightFactor())
+                params.set_alternativeRoutesWeightFactor(alternativeRoutes.getWeightFactor());
+            if (alternativeRoutes.hasShareFactor())
+                params.set_alternativeRoutesShareFactor(alternativeRoutes.getShareFactor());
+        }
+
         params.setConsiderTurnRestrictions(false);
 
         routingRequest.setSearchParameters(params);
@@ -218,6 +231,7 @@ public class RouteRequestHandler extends GenericHandler {
             params.setVehicleType(convertVehicleType(options.getVehicleType(), params.getProfileType()));
         return params;
     }
+
     private  boolean convertIncludeGeometry(RouteRequest request) throws IncompatibleParameterException {
         boolean includeGeometry = request.getIncludeGeometry();
         if(!includeGeometry && request.getResponseType() != APIEnums.RouteResponseType.JSON) {